requirements.txt
numpy~=1.24.1
torch~=2.1.0
Pillow~=9.4.0
wandb~=0.16.6
PyYAML~=6.0.1
scikit-learn~=1.3.2
albumentations~=1.4.6
matplotlib~=3.7.3